+++ NEWS.md
@@ -7,6 +7,8 @@
 
 ### Bug fixes
 
+* Row selection columns now display correctly in tables with column groups
+  ([#52](https://github.com/glin/reactable/issues/52)).
 * `defaultSelected` now works correctly with Crosstalk linked selection.
 * Shiny UI elements in expanded row details are now properly removed when
   collapsed on page changes.